+++ src/components/SecurityContainer/SecurityContainer.js
@@ -1,28 +1,62 @@
 import React, { PureComponent } from 'react'
+import classNames from 'classnames'
 import PropTypes from 'prop-types'
 import Description from '../Description/Description'
+import Indicator from '../Indicator/Indicator'
 import { styles } from './SecurityContainer.styles'
 
 @styles
 export default class SecurityContainer extends PureComponent {
+  constructor (props) {
+    super(props)
+
+    this.onClick = this.onClick.bind(this)
+
+    this.state = {
+      isOpen: false
+    }
+  }
+
   render () {
-    const { id, security, classes } = this.props
+    const { id, security, classes, placedIn } = this.props
     const { name, type, description } = security
     const isSimple = ['apiKey', 'http'].includes(security.type)
 
+    const { isOpen } = this.state
+
+    let indicatorDirection
+    if (isOpen) {
+      indicatorDirection = 'up'
+    } else {
+      indicatorDirection = 'down'
+    }
+
+    let title
+    if (placedIn === 'schema') {
+      title = <h3>{name} <code className={classes.scheme}>type={type}</code></h3>
+    } else if (placedIn === 'method') {
+      title = <h5>{name} <code className={classes.scheme}>type={type}</code></h5>
+    }
+
     return (
-      <section className={classes.securityContainer} id={id}>
-        <h3>{name} <code className={classes.scheme}>type={type}</code></h3>
-        {description && <Description description={description} />}
-        {isSimple && this.renderSimple(security)}
-        {security.type === 'oauth2' && this.renderOAuth2(security)}
-        {security.type === 'openIdConnect' && this.renderOpenIdConnect(security)}
-      </section>
+      <div className={classes.securityContainer}>
+        <div className={classNames(classes.info, classes.isClickable)} onClick={this.onClick}>
+          {title}
+          {<Indicator direction={indicatorDirection} />}
+        </div>
+        <section id={id} >
+          {isOpen && description && <Description isInline description={description} />}
+          {isOpen && isSimple && this.renderSimple(security)}
+          {isOpen && security.type === 'oauth2' && this.renderOAuth2(security)}
+          {isOpen && security.type === 'openIdConnect' && this.renderOpenIdConnect(security)}
+        </section>
+      </div>
     )
   }
 
   renderSimple (security) {
     const { example, bearerFormat } = security
+    const { classes } = this.props
     let usage
 
     if (security.in === 'query') {
@@ -32,24 +66,31 @@ export default class SecurityContainer extends PureComponent {
     }
 
     return (
-      <div>
+      <div className={classes.simple} >
         {usage}
       </div>
     )
   }
 
   renderOAuth2 (security) {
-    const { classes } = this.props
+    const { classes, placedIn } = this.props
     const { flows } = security
 
     return (
       <div>
         {Object.keys(flows).map((flowKey) => {
           const flow = flows[flowKey]
+          let title
+
+          if (placedIn === 'schema') {
+            title = <h4><code>{flowKey}</code> flow</h4>
+          } else if (placedIn === 'method') {
+            title = <h5><code>{flowKey}</code> flow</h5>
+          }
 
           return (
             <div className={classes.flowType} key={flowKey}>
-              <h4><code>{flowKey}</code> flow</h4>
+              {title}
               <dl className={classes.inlinePairs}>
                 {flow.authorizationUrl && [
                   <dt key='auth'>Authorization URL</dt>,
@@ -98,10 +139,19 @@ export default class SecurityContainer extends PureComponent {
       </div>
     )
   }
+
+  onClick () {
+    if (this.state.isOpen) {
+      this.setState({ isOpen: false })
+    } else {
+      this.setState({ isOpen: true })
+    }
+  }
 }
 
 SecurityContainer.propTypes = {
   id: PropTypes.string,
   security: PropTypes.object,
-  classes: PropTypes.object
+  classes: PropTypes.object,
+  placedIn: PropTypes.oneOf(['schema', 'method'])
 }